Create shared test mixin for testing OAuth Scopes and various Auth. ARCH-168
