+++ lms/djangoapps/certificates/apis/v0/tests/test_views.py
@@ -2,16 +2,12 @@
 Tests for the Certificate REST APIs.
 """
 # pylint: disable=missing-docstring
-from datetime import datetime, timedelta
-from enum import Enum
 from itertools import product
 import ddt
-from mock import patch
 
 from django.urls import reverse
 from django.utils import timezone
 from freezegun import freeze_time
-from oauth2_provider import models as dot_models
 from rest_framework import status
 from rest_framework.test import APITestCase
 
@@ -19,32 +15,21 @@ from course_modes.models import CourseMode
 from lms.djangoapps.certificates.apis.v0.views import CertificatesDetailView
 from lms.djangoapps.certificates.models import CertificateStatuses
 from lms.djangoapps.certificates.tests.factories import GeneratedCertificateFactory
-from openedx.core.djangoapps.oauth_dispatch.jwt import _create_jwt
 from openedx.core.djangoapps.oauth_dispatch.toggles import ENFORCE_JWT_SCOPES
+from openedx.core.djangoapps.user_authn.tests.utils import AuthType, AuthAndScopesTestMixin
 from student.tests.factories import UserFactory
 from xmodule.modulestore.tests.django_utils import SharedModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory
 
-USER_PASSWORD = 'test'
-
-
-class AuthType(Enum):
-    session = 1
-    oauth = 2
-    jwt = 3
-    jwt_restricted = 4
-
-
-JWT_AUTH_TYPES = [AuthType.jwt, AuthType.jwt_restricted]
-
 
 @ddt.ddt
-class CertificatesRestApiTest(SharedModuleStoreTestCase, APITestCase):
+class CertificatesRestApiTest(AuthAndScopesTestMixin, SharedModuleStoreTestCase, APITestCase):
     """
     Test for the Certificates REST APIs
     """
     shard = 4
     now = timezone.now()
+    default_scopes = CertificatesDetailView.required_scopes
 
     @classmethod
     def setUpClass(cls):
@@ -62,10 +47,6 @@ class CertificatesRestApiTest(SharedModuleStoreTestCase, APITestCase):
 
         super(CertificatesRestApiTest, self).setUp()
 
-        self.student = UserFactory.create(password=USER_PASSWORD)
-        self.student_no_cert = UserFactory.create(password=USER_PASSWORD)
-        self.staff_user = UserFactory.create(password=USER_PASSWORD, is_staff=True)
-
         GeneratedCertificateFactory.create(
             user=self.student,
             course_id=self.course.id,
@@ -77,7 +58,18 @@ class CertificatesRestApiTest(SharedModuleStoreTestCase, APITestCase):
 
         self.namespaced_url = 'certificates_api:v0:certificates:detail'
 
-    def _assert_certificate_response(self, response):
+    def get_url(self, username):
+        """ This method is required by AuthAndScopesTestMixin. """
+        return reverse(
+            self.namespaced_url,
+            kwargs={
+                'course_id': self.course.id,
+                'username': username
+            }
+        )
+
+    def assert_success_response_for_student(self, response):
+        """ This method is required by AuthAndScopesTestMixin. """
         self.assertEqual(
             response.data,
             {
@@ -92,86 +84,12 @@ class CertificatesRestApiTest(SharedModuleStoreTestCase, APITestCase):
             }
         )
 
-    def _get_url(self, username):
-        """
-        Helper function to create the url for certificates
-        """
-        return reverse(
-            self.namespaced_url,
-            kwargs={
-                'course_id': self.course.id,
-                'username': username
-            }
-        )
-
-    def _create_oauth_token(self, user):
-        dot_app_user = UserFactory.create(password=USER_PASSWORD)
-        dot_app = dot_models.Application.objects.create(
-            name='test app',
-            user=dot_app_user,
-            client_type='confidential',
-            authorization_grant_type='authorization-code',
-            redirect_uris='http://localhost:8079/complete/edxorg/'
-        )
-        return dot_models.AccessToken.objects.create(
-            user=user,
-            application=dot_app,
-            expires=datetime.utcnow() + timedelta(weeks=1),
-            scope='read write',
-            token='test_token',
-        )
-
-    def _create_jwt_token(self, user, auth_type, scopes=None, include_org_filter=True, include_me_filter=False):
-        filters = []
-        if include_org_filter:
-            filters += ['content_org:{}'.format(self.course.id.org)]
-        if include_me_filter:
-            filters += ['user:me']
-
-        if scopes is None:
-            scopes = CertificatesDetailView.required_scopes
-
-        return _create_jwt(
-            user,
-            scopes=scopes,
-            is_restricted=(auth_type == AuthType.jwt_restricted),
-            filters=filters,
-        )
-
-    def _get_response(self, requesting_user, auth_type, url=None, token=None):
-        auth_header = None
-        if auth_type == AuthType.session:
-            self.client.login(username=requesting_user.username, password=USER_PASSWORD)
-        elif auth_type == AuthType.oauth:
-            if not token:
-                token = self._create_oauth_token(requesting_user)
-            auth_header = "Bearer {0}".format(token)
-        else:
-            assert auth_type in JWT_AUTH_TYPES
-            if not token:
-                token = self._create_jwt_token(requesting_user, auth_type)
-            auth_header = "JWT {0}".format(token)
-
-        extra = dict(HTTP_AUTHORIZATION=auth_header) if auth_header else {}
-        return self.client.get(
-            url if url else self._get_url(self.student.username),
-            **extra
-        )
-
-    def _assert_in_log(self, text, mock_log_method):
-        self.assertTrue(mock_log_method.called)
-        self.assertIn(text, mock_log_method.call_args_list[0][0][0])
-
-    def test_anonymous_user(self):
-        resp = self.client.get(self._get_url(self.student.username))
-        self.assertEqual(resp.status_code, status.HTTP_401_UNAUTHORIZED)
-
-    @ddt.data(*list(AuthType))
-    def test_no_certificate(self, auth_type):
-        resp = self._get_response(
-            self.student_no_cert,
-            auth_type,
-            url=self._get_url(self.student_no_cert.username),
+    def test_no_certificate(self):
+        student_no_cert = UserFactory.create(password=self.user_password)
+        resp = self.get_response(
+            AuthType.session,
+            requesting_user=student_no_cert,
+            requested_user=student_no_cert,
         )
         self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)
         self.assertIn('error_code', resp.data)
@@ -179,116 +97,3 @@ class CertificatesRestApiTest(SharedModuleStoreTestCase, APITestCase):
             resp.data['error_code'],
             'no_certificate_for_user',
         )
-
-    @ddt.data(*product(list(AuthType), (True, False)))
-    @ddt.unpack
-    def test_self_user(self, auth_type, scopes_enforced):
-        with ENFORCE_JWT_SCOPES.override(active=scopes_enforced):
-            resp = self._get_response(self.student, auth_type)
-            self.assertEqual(resp.status_code, status.HTTP_200_OK)
-            self._assert_certificate_response(resp)
-
-    @ddt.data(*product(list(AuthType), (True, False)))
-    @ddt.unpack
-    def test_inactive_user(self, auth_type, scopes_enforced):
-        with ENFORCE_JWT_SCOPES.override(active=scopes_enforced):
-            self.student.is_active = False
-            self.student.save()
-
-            resp = self._get_response(self.student, auth_type)
-            self.assertEqual(resp.status_code, status.HTTP_200_OK)
-
-    @ddt.data(*product(list(AuthType), (True, False)))
-    @ddt.unpack
-    def test_staff_user(self, auth_type, scopes_enforced):
-        with ENFORCE_JWT_SCOPES.override(active=scopes_enforced):
-            resp = self._get_response(self.staff_user, auth_type)
-            self.assertEqual(resp.status_code, status.HTTP_200_OK)
-
-    @patch('edx_rest_framework_extensions.permissions.log')
-    @ddt.data(*product(list(AuthType), (True, False)))
-    @ddt.unpack
-    def test_another_user(self, auth_type, scopes_enforced, mock_log):
-        """ Returns 403 for OAuth and Session auth with IsUserInUrl. """
-        with ENFORCE_JWT_SCOPES.override(active=scopes_enforced):
-            resp = self._get_response(self.student_no_cert, auth_type)
-
-            # Restricted JWT tokens without the user:me filter have access to other users
-            expected_jwt_access_granted = scopes_enforced and auth_type == AuthType.jwt_restricted
-
-            self.assertEqual(
-                resp.status_code,
-                status.HTTP_200_OK if expected_jwt_access_granted else status.HTTP_403_FORBIDDEN,
-            )
-            if not expected_jwt_access_granted:
-                self._assert_in_log("IsUserInUrl", mock_log.info)
-
-    @patch('edx_rest_framework_extensions.permissions.log')
-    @ddt.data(*product(JWT_AUTH_TYPES, (True, False)))
-    @ddt.unpack
-    def test_jwt_no_scopes(self, auth_type, scopes_enforced, mock_log):
-        """ Returns 403 when scopes are enforced with JwtHasScope. """
-        with ENFORCE_JWT_SCOPES.override(active=scopes_enforced):
-            jwt_token = self._create_jwt_token(self.student, auth_type, scopes=[])
-            resp = self._get_response(self.student, AuthType.jwt, token=jwt_token)
-
-            is_enforced = scopes_enforced and auth_type == AuthType.jwt_restricted
-            self.assertEqual(resp.status_code, status.HTTP_403_FORBIDDEN if is_enforced else status.HTTP_200_OK)
-
-            if is_enforced:
-                self._assert_in_log("JwtHasScope", mock_log.warning)
-
-    @patch('edx_rest_framework_extensions.permissions.log')
-    @ddt.data(*product(JWT_AUTH_TYPES, (True, False)))
-    @ddt.unpack
-    def test_jwt_no_filter(self, auth_type, scopes_enforced, mock_log):
-        """ Returns 403 when scopes are enforced with JwtHasContentOrgFilterForRequestedCourse. """
-        with ENFORCE_JWT_SCOPES.override(active=scopes_enforced):
-            jwt_token = self._create_jwt_token(self.student, auth_type, include_org_filter=False)
-            resp = self._get_response(self.student, AuthType.jwt, token=jwt_token)
-
-            is_enforced = scopes_enforced and auth_type == AuthType.jwt_restricted
-            self.assertEqual(resp.status_code, status.HTTP_403_FORBIDDEN if is_enforced else status.HTTP_200_OK)
-
-            if is_enforced:
-                self._assert_in_log("JwtHasContentOrgFilterForRequestedCourse", mock_log.warning)
-
-    @ddt.data(*product(JWT_AUTH_TYPES, (True, False)))
-    @ddt.unpack
-    def test_jwt_on_behalf_of_user(self, auth_type, scopes_enforced):
-        with ENFORCE_JWT_SCOPES.override(active=scopes_enforced):
-            jwt_token = self._create_jwt_token(self.student, auth_type, include_me_filter=True)
-
-            resp = self._get_response(self.student, AuthType.jwt, token=jwt_token)
-            self.assertEqual(resp.status_code, status.HTTP_200_OK)
-
-    @patch('edx_rest_framework_extensions.permissions.log')
-    @ddt.data(*product(JWT_AUTH_TYPES, (True, False)))
-    @ddt.unpack
-    def test_jwt_on_behalf_of_other_user(self, auth_type, scopes_enforced, mock_log):
-        """ Returns 403 when scopes are enforced with JwtHasUserFilterForRequestedUser. """
-        with ENFORCE_JWT_SCOPES.override(active=scopes_enforced):
-            jwt_token = self._create_jwt_token(self.student_no_cert, auth_type, include_me_filter=True)
-            resp = self._get_response(self.student, AuthType.jwt, token=jwt_token)
-
-            self.assertEqual(resp.status_code, status.HTTP_403_FORBIDDEN)
-
-            if scopes_enforced and auth_type == AuthType.jwt_restricted:
-                self._assert_in_log("JwtHasUserFilterForRequestedUser", mock_log.warning)
-            else:
-                self._assert_in_log("IsUserInUrl", mock_log.info)
-
-    def test_valid_oauth_token(self):
-        resp = self._get_response(self.student, AuthType.oauth)
-        self.assertEqual(resp.status_code, status.HTTP_200_OK)
-
-    def test_invalid_oauth_token(self):
-        resp = self._get_response(self.student, AuthType.oauth, token="fooooooooooToken")
-        self.assertEqual(resp.status_code, status.HTTP_401_UNAUTHORIZED)
-
-    def test_expired_oauth_token(self):
-        token = self._create_oauth_token(self.student)
-        token.expires = datetime.utcnow() - timedelta(weeks=1)
-        token.save()
-        resp = self._get_response(self.student, AuthType.oauth, token=token)
-        self.assertEqual(resp.status_code, status.HTTP_401_UNAUTHORIZED)
