+++ lms/djangoapps/grades/api/v1/tests/test_views.py
@@ -15,11 +15,13 @@ from rest_framework.test import APITestCase
 from six import text_type
 
 from lms.djangoapps.courseware.tests.factories import GlobalStaffFactory, StaffFactory
+from lms.djangoapps.grades.api.v1.views import CourseGradesView
 from lms.djangoapps.grades.config.waffle import waffle_flags, WRITABLE_GRADEBOOK
 from lms.djangoapps.grades.course_data import CourseData
 from lms.djangoapps.grades.course_grade import CourseGrade
 from lms.djangoapps.grades.subsection_grade import ReadSubsectionGrade
 from openedx.core.djangoapps.content.course_overviews.tests.factories import CourseOverviewFactory
+from openedx.core.djangoapps.user_authn.tests.utils import AuthAndScopesTestMixin
 from openedx.core.djangoapps.waffle_utils.testutils import override_waffle_flag
 from student.tests.factories import CourseEnrollmentFactory, UserFactory
 from xmodule.modulestore.tests.factories import CourseFactory, ItemFactory
@@ -79,27 +81,22 @@ class GradeViewTestMixin(SharedModuleStoreTestCase):
         )
         cls.course_key = cls.course.id
 
-        cls.password = 'test'
-        cls.student = UserFactory(username='dummy', password=cls.password)
-        cls.other_student = UserFactory(username='foo', password=cls.password)
-        cls.other_user = UserFactory(username='bar', password=cls.password)
-        cls.staff = StaffFactory(course_key=cls.course_key, password=cls.password)
-        cls.global_staff = GlobalStaffFactory.create()
-        cls._create_user_enrollments(cls.course, cls.student, cls.other_student)
-
-    def setUp(self):
-        super(GradeViewTestMixin, self).setUp()
-        self.client.login(username=self.student.username, password=self.password)
-
-    @classmethod
-    def _create_user_enrollments(cls, course, *users):
+    def _create_user_enrollments(self, *users):
         date = datetime(2013, 1, 22, tzinfo=UTC)
         for user in users:
             CourseEnrollmentFactory(
-                course_id=course.id,
+                course_id=self.course.id,
                 user=user,
                 created=date,
             )
+
+    def setUp(self):
+        super(GradeViewTestMixin, self).setUp()
+        self.password = 'test'
+        self.global_staff = GlobalStaffFactory.create()
+        self.student = UserFactory(password=self.password)
+        self.other_student = UserFactory(password=self.password)
+        self._create_user_enrollments(self.student, self.other_student)
 
     @classmethod
     def _create_test_course_with_default_grading_policy(cls, display_name, run):
@@ -141,12 +138,13 @@ class GradeViewTestMixin(SharedModuleStoreTestCase):
 
 
 @ddt.ddt
-class SingleUserGradesTests(GradeViewTestMixin, APITestCase):
+class SingleUserGradesTests(GradeViewTestMixin, AuthAndScopesTestMixin, APITestCase):
     """
     Tests for grades related to a course and specific user
         e.g. /api/grades/v1/courses/{course_id}/?username={username}
              /api/grades/v1/courses/?course_id={course_id}&username={username}
     """
+    default_scopes = CourseGradesView.required_scopes
 
     @classmethod
     def setUpClass(cls):
@@ -154,9 +152,7 @@ class SingleUserGradesTests(GradeViewTestMixin, APITestCase):
         cls.namespaced_url = 'grades_api:v1:course_grades'
 
     def get_url(self, username):
-        """
-        Helper function to create the url
-        """
+        """ This method is required by AuthAndScopesTestMixin. """
         base_url = reverse(
             self.namespaced_url,
             kwargs={
@@ -165,38 +161,25 @@ class SingleUserGradesTests(GradeViewTestMixin, APITestCase):
         )
         return "{0}?username={1}".format(base_url, username)
 
-    def test_anonymous(self):
-        """
-        Test that an anonymous user cannot access the API and an error is received.
-        """
-        self.client.logout()
-        resp = self.client.get(self.get_url(self.student.username))
-        self.assertEqual(resp.status_code, status.HTTP_401_UNAUTHORIZED)
-
-    def test_self_get_grade(self):
-        """
-        Test that a user can successfully request her own grade.
-        """
-        resp = self.client.get(self.get_url(self.student.username))
-        self.assertEqual(resp.status_code, status.HTTP_200_OK)
+    def assert_success_response_for_student(self, response):
+        """ This method is required by AuthAndScopesTestMixin. """
+        expected_data = [{
+            'username': self.student.username,
+            'email': self.student.email,
+            'letter_grade': None,
+            'percent': 0.0,
+            'course_id': str(self.course_key),
+            'passed': False
+        }]
+        self.assertEqual(response.data, expected_data)
 
     def test_nonexistent_user(self):
         """
         Test that a request for a nonexistent username returns an error.
         """
-        self.client.logout()
         self.client.login(username=self.global_staff.username, password=self.password)
         resp = self.client.get(self.get_url('IDoNotExist'))
         self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)
-
-    def test_other_get_grade(self):
-        """
-        Test that if a user requests the grade for another user, she receives an error.
-        """
-        self.client.logout()
-        self.client.login(username=self.other_student.username, password=self.password)
-        resp = self.client.get(self.get_url(self.student.username))
-        self.assertEqual(resp.status_code, status.HTTP_403_FORBIDDEN)
 
     def test_self_get_grade_not_enrolled(self):
         """
@@ -204,9 +187,9 @@ class SingleUserGradesTests(GradeViewTestMixin, APITestCase):
         her own grade in a course where she is not enrolled.
         """
         # a user not enrolled in the course cannot request her grade
-        self.client.logout()
-        self.client.login(username=self.other_user.username, password=self.password)
-        resp = self.client.get(self.get_url(self.other_user.username))
+        unenrolled_user = UserFactory(password=self.password)
+        self.client.login(username=unenrolled_user.username, password=self.password)
+        resp = self.client.get(self.get_url(unenrolled_user.username))
         self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)
         self.assertIn('error_code', resp.data)
         self.assertEqual(
@@ -218,6 +201,7 @@ class SingleUserGradesTests(GradeViewTestMixin, APITestCase):
         """
         Test the grade for a user who has not answered any test.
         """
+        self.client.login(username=self.student.username, password=self.password)
         resp = self.client.get(self.get_url(self.student.username))
         self.assertEqual(resp.status_code, status.HTTP_200_OK)
         expected_data = [{
@@ -239,6 +223,7 @@ class SingleUserGradesTests(GradeViewTestMixin, APITestCase):
             """Mocked function to always raise an exception"""
             raise InvalidKeyError('foo', 'bar')
 
+        self.client.login(username=self.student.username, password=self.password)
         with patch('opaque_keys.edx.keys.CourseKey.from_string', side_effect=mock_from_string):
             resp = self.client.get(self.get_url(self.student.username))
 
@@ -253,6 +238,7 @@ class SingleUserGradesTests(GradeViewTestMixin, APITestCase):
         """
         Test that requesting a valid, nonexistent course key returns an error as expected.
         """
+        self.client.login(username=self.student.username, password=self.password)
         base_url = reverse(
             self.namespaced_url,
             kwargs={
@@ -276,6 +262,7 @@ class SingleUserGradesTests(GradeViewTestMixin, APITestCase):
         """
         Test that the user gets her grade in case she answered tests with an insufficient score.
         """
+        self.client.login(username=self.student.username, password=self.password)
         with patch('lms.djangoapps.grades.course_grade_factory.CourseGradeFactory.read') as mock_grade:
             grade_fields = {
                 'letter_grade': grade['letter_grade'],
@@ -295,24 +282,6 @@ class SingleUserGradesTests(GradeViewTestMixin, APITestCase):
 
         expected_data.update(grade)
         self.assertEqual(resp.data, [expected_data])
-
-    def test_staff_can_see_student(self):
-        """
-        Ensure that staff members can see her student's grades.
-        """
-        self.client.logout()
-        self.client.login(username=self.global_staff.username, password=self.password)
-        resp = self.client.get(self.get_url(self.student.username))
-        self.assertEqual(resp.status_code, status.HTTP_200_OK)
-        expected_data = [{
-            'username': self.student.username,
-            'email': self.student.email,
-            'letter_grade': None,
-            'percent': 0.0,
-            'course_id': str(self.course_key),
-            'passed': False
-        }]
-        self.assertEqual(resp.data, expected_data)
 
 
 @ddt.ddt
@@ -342,16 +311,15 @@ class CourseGradesViewTest(GradeViewTestMixin, APITestCase):
         return base_url
 
     def test_anonymous(self):
-        self.client.logout()
         resp = self.client.get(self.get_url())
         self.assertEqual(resp.status_code, status.HTTP_401_UNAUTHORIZED)
 
     def test_student(self):
+        self.client.login(username=self.student.username, password=self.password)
         resp = self.client.get(self.get_url())
         self.assertEqual(resp.status_code, status.HTTP_403_FORBIDDEN)
 
     def test_course_does_not_exist(self):
-        self.client.logout()
         self.client.login(username=self.global_staff.username, password=self.password)
         resp = self.client.get(
             self.get_url(course_key='course-v1:MITx+8.MechCX+2014_T1')
@@ -359,7 +327,6 @@ class CourseGradesViewTest(GradeViewTestMixin, APITestCase):
         self.assertEqual(resp.status_code, status.HTTP_404_NOT_FOUND)
 
     def test_course_no_enrollments(self):
-        self.client.logout()
         self.client.login(username=self.global_staff.username, password=self.password)
         resp = self.client.get(
             self.get_url(course_key=self.empty_course.id)
@@ -368,7 +335,6 @@ class CourseGradesViewTest(GradeViewTestMixin, APITestCase):
         self.assertEqual(resp.data, [])
 
     def test_staff_can_get_all_grades(self):
-        self.client.logout()
         self.client.login(username=self.global_staff.username, password=self.password)
         resp = self.client.get(self.get_url())
 
@@ -409,10 +375,6 @@ class GradebookViewTest(GradeViewTestMixin, APITestCase):
 
         cls.course = CourseFactory.create(display_name='test-course', run='run-1')
         cls.course_overview = CourseOverviewFactory.create(id=cls.course.id)
-
-        # we re-assign cls.course from what's created in the parent class, so we have to
-        # re-create the enrollments, too.
-        cls._create_user_enrollments(cls.course, cls.student, cls.other_student)
 
         cls.chapter_1 = ItemFactory.create(
             category='chapter',
@@ -538,7 +500,6 @@ class GradebookViewTest(GradeViewTestMixin, APITestCase):
         Helper function to login the global staff user, who has permissions to read from the
         Gradebook API.
         """
-        self.client.logout()
         self.client.login(username=self.global_staff.username, password=self.password)
 
     def expected_subsection_grades(self, letter_grade=None):
@@ -637,7 +598,6 @@ class GradebookViewTest(GradeViewTestMixin, APITestCase):
         ]
 
     def test_feature_not_enabled(self):
-        self.client.logout()
         self.client.login(username=self.global_staff.username, password=self.password)
         with override_waffle_flag(self.waffle_flag, active=False):
             resp = self.client.get(
@@ -646,12 +606,12 @@ class GradebookViewTest(GradeViewTestMixin, APITestCase):
             self.assertEqual(status.HTTP_403_FORBIDDEN, resp.status_code)
 
     def test_anonymous(self):
-        self.client.logout()
         with override_waffle_flag(self.waffle_flag, active=True):
             resp = self.client.get(self.get_url())
             self.assertEqual(status.HTTP_401_UNAUTHORIZED, resp.status_code)
 
     def test_student(self):
+        self.client.login(username=self.student.username, password=self.password)
         with override_waffle_flag(self.waffle_flag, active=True):
             resp = self.client.get(self.get_url())
             self.assertEqual(status.HTTP_403_FORBIDDEN, resp.status_code)
